fix connection failure when client on host connect to replicaSet
mongo client on host was not able to reach the replicaSet, even
though it was able to connect to every single node. To solve this,
replica set configuration must be based on the host, and not
on the primary node. In this case, replacing the mongod nodes
hostnames with "localhost" is enough.